requirements.txt
# Versiones compatibles verificadas (marzo 2026)
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
openai>=1.66.0
anthropic>=0.84.0  # CORREGIDO: versión máxima actual
google-generativeai>=0.8.0
pypdf>=5.4.0
chromadb>=0.6.0
tiktoken>=0.9.0
langchain-text-splitters
sv-ttk
langchain-chroma
pyinstaller